app/requirements-dev.txt
-r requirements.txt
pytest==8.3.2
pytest-flask==1.3.0
selenium==4.23.1
webdriver-manager==4.0.2
